Popup  Boxes (depth, bottom time, surface int)
Redid the popup boxes for depth, bottom time, and surface interval to
allow users to enter information instead of a drop down menu. Also did
various checking for valid inputs (positive integers, actual numbers).
If error was found, user was told of what kind of error happened with
the use of message boxes